+++ ImagineProject/Controllers/RecintoController.cs
@@ -14,6 +14,21 @@ namespace ImagineProject.Controllers
     {
         private Db_ImagineEntities db = new Db_ImagineEntities();
 
+        public bool HaveReferences(int id)
+        {
+            bool resultado = false;
+            var cant = (db.RecintoPorticos.Where(rp => rp.id_recinto == id)).Count();
+            if (cant > 0)
+            {
+                resultado = true;
+            }
+            else if (cant == 0)
+            {
+                resultado = false;
+            }
+            return resultado;
+        }       
+
         //
         // GET: /Recinto/
 
@@ -106,11 +121,22 @@ namespace ImagineProject.Controllers
 
         [HttpPost, ActionName("Delete")]
         public ActionResult DeleteConfirmed(int id)
-        {            
-            Recinto recinto = db.Recintos.Find(id);
-            db.Recintos.Remove(recinto);
-            db.SaveChanges();
-            return RedirectToAction("Index");
+        {
+            if (!HaveReferences(id))
+            {
+                Recinto recinto = db.Recintos.Find(id);
+                db.Recintos.Remove(recinto);
+                db.SaveChanges();
+                return RedirectToAction("Index");
+            }
+            else
+            {
+                Error error = new Error();
+                error.Message = "Error: No puede eliminar este recinto porque tiene porticos asociados.";
+                error.Action = "Delete";
+                error.Controller = "Recinto";
+                return View("~/Views/Shared/Error.aspx", error);
+            }
         }
 
         protected override void Dispose(bool disposing)
