+++ ImagineProject/Controllers/ZonaController.cs
@@ -14,6 +14,21 @@ namespace ImagineProject.Controllers
     {
         private Db_ImagineEntities db = new Db_ImagineEntities();
 
+        public bool HaveReferences(int id)
+        {
+            bool resultado = false;
+            var cant = (db.ZonaPaises.Where(zp => zp.id_zona == id)).Count();
+            if (cant > 0)
+            {
+                resultado = true;
+            }
+            else if (cant == 0)
+            {
+                resultado = false;
+            }
+            return resultado;
+        }  
+
         //
         // GET: /Zona/
 
@@ -99,10 +114,21 @@ namespace ImagineProject.Controllers
         [HttpPost, ActionName("Delete")]
         public ActionResult DeleteConfirmed(int id)
         {
-            Zona zona = db.Zonas.Find(id);
-            db.Zonas.Remove(zona);
-            db.SaveChanges();
-            return RedirectToAction("Index");
+            if (!HaveReferences(id))
+            {
+                Zona zona = db.Zonas.Find(id);
+                db.Zonas.Remove(zona);
+                db.SaveChanges();
+                return RedirectToAction("Index");
+            }
+            else
+            {
+                Error error = new Error();
+                error.Message = "Error: No puede eliminar esta zona porque tiene países asociados.";
+                error.Action = "Delete";
+                error.Controller = "Zona";
+                return View("~/Views/Shared/Error.aspx", error);
+            }
         }
 
         protected override void Dispose(bool disposing)
